Edtech-Website-main/requirements.txt
streamlit
streamlit_lottie
pillow
requests